pyproject.toml
[project]
name = "llm-router"
version = "0.1.0"
description = "LLM Router — intelligent model routing proxy"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.30.0",
    "ruamel.yaml>=0.18",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["llm_router*"]
